Improves the caching of `AutoCompleteSystem._project_users and `SimplifiedPermissions.get_permission_actions`
Uses tracs `@cached` annotation for the expensive method
`AutoCompleteSystem._project_users`.  I didn't believe in it at first
because it makes database calls, but after reading the docs I think it
sounds pretty good.  A side effect is that the method is changed into a
property, but that doesn't matter.

Moves the cache variables for
`SimplifiedPermissions.get_permission_actions` from the class to the
instance.  We had a discussion on xmpp wether the classes are shared
between define projects or not.  While we didn't come to a conclusion I
think they most likely are, so making them instance variables is safer.
I did not use `@cached` for this since it's an api method and cannot be
a property.  The method could of course just access a cached property,
but I think this will do just fine.

Refs #3723

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@21241 484df14f-75d2-4b3f-8ba0-490e05069a01