Escapes strings that will be injected into the page on the backlog page
Separates the ticket fields into two categories: `safe` and `unsafe`.
Fields that don't allow users to input text are considered `safe` and
are sent to the front end unescaped.  All other fields can contain
injected javascript.

Note that this ticket does not fully render the backlog page harmless.
One can still create a milestone with javascript in the name which makes
the milestone dropdown dangerous.  This happens on all pages with
select2 dialogs, not just the backlog.

Refs #4972

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@24719 484df14f-75d2-4b3f-8ba0-490e05069a01