Fixes #5087 - Adds ability to jump by larger positions in the backlog.
Three new icons now appear when you hover over a backlog milestone ticket:
 * Move to top of current milestone
 * Move to bottom of current milestone
 * More to a custom milestone / custom position

The custom options are displayed inside a new dialog. This includes a select2 widget
listing milestone options, and a input field with some HTML5 attributes such as a
min and max provided to add some basic validation when using the step widget.

You can of course enter an invalid position, but this will be caught in the
manually_move_tickets() function and another dialog rendered to explain the problem.

One problem I've had to overcome is the need for a milestone to be rendered in order
to use the jQuery sortable plugin. This means that if a milestone is selected in the
more options dialog which is not already displayed, we call the add_milestone() method.
If four milestones are already shown, we display the new milestoneFailDialog.

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@24960 484df14f-75d2-4b3f-8ba0-490e05069a01