#5087 - Added a comment attempting to explain why we check the current milestone and position.
For example: Suppose we want to move ticket A, currently at position 1, to position 3 in the same milestone, currently held by ticket C.
We must insert ticket A after ticket C in this case, because ticket Bs position of 3 will fall to position 2 at the time of the swap. This
will leave ticket A correctly positioned at 3.

We don't need to account for this when moving to a new milestone, because the ticket currently at the position we wish to move to, will not
have its position changed when we move the ticket across.

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@24992 484df14f-75d2-4b3f-8ba0-490e05069a01